+++ lib/twenv/command_mixins/tweet_io.rb
@@ -1,15 +1,9 @@
 module TWEnv::Command::TweetIO
   require 'json'
 
-  class Tweet < OpenStruct
-    def to_json(options = {})
-      @table.to_json(options)
-    end
-  end
-
   def read_tweets_array(path)
     tweets = JSON.parse File.read(path)
-    tweets.map {|tweet| Tweet.new(tweet)}
+    tweets.map {|tweet| TWEnv::Struct.from_hash(tweet)}
   end
 
   def write_tweets_array(path, tweets)